+++ cartographer/mapping/internal/constraints/constraint_builder_2d.cc
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "cartographer/mapping/internal/2d/pose_graph/constraint_builder_2d.h"
+#include "cartographer/mapping/internal/constraints/constraint_builder_2d.h"
 
 #include <cmath>
 #include <functional>
@@ -39,7 +39,7 @@
 
 namespace cartographer {
 namespace mapping {
-namespace pose_graph {
+namespace constraints {
 
 static auto* kConstraintsSearchedMetric = metrics::Counter::Null();
 static auto* kConstraintsFoundMetric = metrics::Counter::Null();
@@ -54,7 +54,7 @@ transform::Rigid2d ComputeSubmapPose(const Submap2D& submap) {
 }
 
 ConstraintBuilder2D::ConstraintBuilder2D(
-    const pose_graph::proto::ConstraintBuilderOptions& options,
+    const constraints::proto::ConstraintBuilderOptions& options,
     common::ThreadPoolInterface* const thread_pool)
     : options_(options),
       thread_pool_(thread_pool),
@@ -328,6 +328,6 @@ void ConstraintBuilder2D::RegisterMetrics(metrics::FamilyFactory* factory) {
   kGlobalConstraintScoresMetric = scores->Add({{"search_region", "global"}});
 }
 
-}  // namespace pose_graph
+}  // namespace constraints
 }  // namespace mapping
 }  // namespace cartographer
