+++ cartographer/mapping/pose_graph/proto/constraint_builder_options.proto
@@ -14,7 +14,7 @@
 
 syntax = "proto3";
 
-package cartographer.mapping.pose_graph.proto;
+package cartographer.mapping.constraints.proto;
 
 import "cartographer/mapping/2d/scan_matching/proto/ceres_scan_matcher_options_2d.proto";
 import "cartographer/mapping/2d/scan_matching/proto/fast_correlative_scan_matcher_options_2d.proto";
