+++ cartographer/metrics/register.cc
@@ -17,17 +17,17 @@
 #include "cartographer/metrics/register.h"
 
 #include "cartographer/mapping/internal/2d/local_trajectory_builder_2d.h"
-#include "cartographer/mapping/internal/2d/pose_graph/constraint_builder_2d.h"
 #include "cartographer/mapping/internal/3d/local_trajectory_builder_3d.h"
-#include "cartographer/mapping/internal/3d/pose_graph/constraint_builder_3d.h"
+#include "cartographer/mapping/internal/constraints/constraint_builder_2d.h"
+#include "cartographer/mapping/internal/constraints/constraint_builder_3d.h"
 #include "cartographer/mapping/internal/global_trajectory_builder.h"
 
 namespace cartographer {
 namespace metrics {
 
 void RegisterAllMetrics(FamilyFactory* registry) {
-  mapping::pose_graph::ConstraintBuilder2D::RegisterMetrics(registry);
-  mapping::pose_graph::ConstraintBuilder3D::RegisterMetrics(registry);
+  mapping::constraints::ConstraintBuilder2D::RegisterMetrics(registry);
+  mapping::constraints::ConstraintBuilder3D::RegisterMetrics(registry);
   mapping::GlobalTrajectoryBuilderRegisterMetrics(registry);
   mapping::LocalTrajectoryBuilder2D::RegisterMetrics(registry);
   mapping::LocalTrajectoryBuilder3D::RegisterMetrics(registry);
